# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Learning Rest API DOCUMENTATION",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dashboard/teachers": {
            "get": {
                "tags": [
                    "Dashboard Api Teachers"
                ],
                "summary": "get all teachers",
                "operationId": "095c9eacc2d4a574d2558effda4e1121",
                "parameters": [
                    {
                        "name": "publish",
                        "in": "query",
                        "description": "filter teachers with publish (publish , unpublish)",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "filter search name , email or phone",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dashboard Api Teachers"
                ],
                "summary": "Add New Teacher",
                "operationId": "a4174eb70d4cc7447bfdac5c59366e17",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "image": {
                                        "type": "file",
                                        "example": "path image"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/dashboard/teachers/{teacher_id}": {
            "get": {
                "tags": [
                    "Dashboard Api Teachers"
                ],
                "summary": "show teacher",
                "operationId": "37091fd84dd80a7ab3dc59f7fd882f30",
                "parameters": [
                    {
                        "name": "teacher_id",
                        "in": "path",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Dashboard Api Teachers"
                ],
                "summary": "update teacher",
                "operationId": "6b64ed9a5fdc77c6d03d5733cd8cdf8e",
                "parameters": [
                    {
                        "name": "teacher_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "publish": {
                                        "type": "boolen",
                                        "example": "integer"
                                    },
                                    "permissions": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "user_create"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Dashboard Api Teachers"
                ],
                "summary": "Delete Teacher",
                "operationId": "bde696a6336081d023c9fec539c849c1",
                "parameters": [
                    {
                        "name": "teacher_id",
                        "in": "path",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/dashboard/teachers/{teacher_id}/approve": {
            "put": {
                "tags": [
                    "Dashboard Api Teachers"
                ],
                "summary": "Approve Teachers",
                "operationId": "ac592cea25e22757e40d1b5d64f86af7",
                "parameters": [
                    {
                        "name": "teacher_id",
                        "in": "path",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "publish": {
                                        "type": "boolen",
                                        "example": "publish or unpublish"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}